Added owner property to car model and accessory model. Template changes depending if there is an user and if the user is owner.

+++ Workshop - Session and Authentication/services/cars.js
@@ -30,24 +30,45 @@ async function createCar(car) {
     await Car.create(car);
 }
 
-async function deleteById(id) {
+async function deleteById(id, ownerId) {
+    const car = await Car.findById(id);
+
+    if (car.owner != ownerId) {
+        return false;
+    }
     await Car.findByIdAndDelete(id);
+
+    return true;
 }
 
-async function updateById(id, car) {
+async function updateById(id, car, ownerId) {
+
     const existing = await Car.findById(id);
+
+    if (existing.owner != ownerId) {
+        return false;
+    }
+
     existing.name = car.name;
     existing.description = car.description;
     existing.imgUrl = car.imgUrl || undefined;
     existing.price = Number(car.price);
     existing.accessories = car.accessories;
+
     await existing.save();
+
+    return true;
 }
 
-async function attachAccessory(carId, accessoryId) {
+async function attachAccessory(carId, accessoryId, ownerId) {
     const existing = await Car.findById(carId);
+    
+    if (existing.owner != ownerId) return false;
+
     existing.accessories.push(accessoryId);
     await existing.save();
+
+    return true;
 }
 
 module.exports = () => (req, res, next) => {
